fix: resolve all TypeScript and Go test compilation errors (20 → 0 errors)
TypeScript Fixes:
- Restore 5 missing runtime modules (generic-{child-supervisor,readiness,supervisor-errors,lifecycle-wait,runtime-state}.ts) from commit 6fb9b8b
- Export knowledge/types from service package for UI access
- Add type annotations to 6 implicit any parameters in knowledge-graph-view.ts
- Add exhaustive switch statements with never assertions in knowledge-panel.ts and knowledge-settings.ts
- Add missing network_access permission labels in permission-modal.ts
- Update service-client.ts to use proper KnowledgeStatusView types instead of inline objects

Go Test Fixes:
- Remove unused database/sql import from llm_config_test.go
- Fix jwtAuth.GenerateToken calls: 1 arg → 3 args (userID, email, expiresIn) in llm_config_test.go
- Fix float32/float64 type mismatch in embedding_performance_test.go (cast minSimilarity to float32)
- Add proper error checking in import_test.go (replace _ with require.NoError)

All TypeScript typecheck now passes (tsc -b). Go test compilation errors resolved.

+++ app/backend/tests/integration/embedding_performance_test.go
@@ -121,7 +121,7 @@ func TestEmbeddingConsistency(t *testing.T) {
 	similarity := cosineSimilarity(vecs1[0], vecs2[0])
 	t.Logf("Embedding consistency (cosine similarity): %.6f", similarity)
 
-	minSimilarity := 0.99
+	minSimilarity := float32(0.99)
 	if similarity < minSimilarity {
 		t.Logf("WARNING: Embeddings not perfectly consistent (similarity: %.6f, expected >= %.2f)", similarity, minSimilarity)
 	}
